Former-commit-id: 64720936105e7699b4fc40ad56438084f0cfc1bf

+++ hpz/hpz/frs/upload_service.py
@@ -1,6 +1,7 @@
 """
 This module describes the file upload endpoint for HPZ.
 """
+import logging
 
 __author__ = 'ablum,'
 __author__ = 'tshewchuk'
@@ -13,6 +14,8 @@ from pyramid.view import view_config
 
 from hpz.database.file_registry import FileRegistry
 
+logger = logging.getLogger(__name__)
+
 
 @view_config(route_name='files', renderer='json', request_method='POST')
 def file_upload_service(context, request):
@@ -21,13 +24,19 @@ def file_upload_service(context, request):
     file_name = request.headers['Filename']
     base_upload_path = request.registry.settings['hpz.frs.upload_base_path']
 
-    file_pathname = os.path.join(base_upload_path, registration_id + '__' + file_name)
+    if FileRegistry.is_file_registered(registration_id):
 
-    input_file = request.POST['file'].file
+        file_pathname = os.path.join(base_upload_path, registration_id + '__' + file_name)
 
-    with open(file_pathname, mode='wb') as output_file:
-        shutil.copyfileobj(input_file, output_file)
+        input_file = request.POST['file'].file
 
-    FileRegistry.file_upload_request(registration_id, file_pathname)
+        with open(file_pathname, mode='wb') as output_file:
+            shutil.copyfileobj(input_file, output_file)
+
+        FileRegistry.file_upload_request(registration_id, file_pathname)
+        logger.error('This file was successfully uploaded')
+
+    else:
+        logger.error('The file attempting to be upload is not registered')
 
     return Response()
